feat: EntityObject 제거 및 EntityClass에 모두 이관
