+++ src/test/java/hibernate/dml/DeleteQueryBuilderTest.java
@@ -1,7 +1,6 @@
 package hibernate.dml;
 
 import hibernate.entity.EntityClass;
-import hibernate.entity.EntityObject;
 import jakarta.persistence.*;
 import org.junit.jupiter.api.Test;
 
@@ -15,10 +14,10 @@ class DeleteQueryBuilderTest {
     void delete쿼리를_생성한다() {
         // given
         String expected = "delete from test_entity where id = 1;";
-        EntityObject<TestEntity> givenEntityObject = new EntityObject<>(new EntityClass<>(TestEntity.class), new TestEntity(1L, "최진영", "jinyoungchoi95@gmail.com"));
+        TestEntity givenEntity = new TestEntity(1L, "최진영", "jinyoungchoi95@gmail.com");
 
         // when
-        String actual = deleteQueryBuilder.generateQuery(givenEntityObject)
+        String actual = deleteQueryBuilder.generateQuery(new EntityClass<>(TestEntity.class), givenEntity.id)
                 .toLowerCase();
 
         // then
