+++ src/test/java/hibernate/dml/InsertQueryBuilderTest.java
@@ -1,7 +1,6 @@
 package hibernate.dml;
 
 import hibernate.entity.EntityClass;
-import hibernate.entity.EntityObject;
 import jakarta.persistence.*;
 import org.junit.jupiter.api.Test;
 
@@ -18,7 +17,7 @@ class InsertQueryBuilderTest {
         String expected = "insert into test_entity (id, nick_name) values (1, '최진영');";
 
         // when
-        String actual = insertQueryBuilder.generateQuery(new EntityObject<>(new EntityClass<>(TestEntity.class), givenEntity))
+        String actual = insertQueryBuilder.generateQuery(new EntityClass<>(TestEntity.class), givenEntity)
                 .toLowerCase();
 
         // then
