+++ src/test/java/hibernate/entity/EntityClassTest.java
@@ -1,14 +1,14 @@
 package hibernate.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import hibernate.entity.column.EntityColumn;
+import jakarta.persistence.*;
 import org.junit.jupiter.api.Test;
 
-import java.lang.reflect.InvocationTargetException;
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 class EntityClassTest {
 
@@ -44,6 +44,29 @@ class EntityClassTest {
                 .hasMessage("기본 생성자가 존재하지 않습니다.");
     }
 
+    @Test
+    void 필드명과_데이터가_담긴_Map을_반환한다() {
+        TestEntity givenObject = new TestEntity(1L, "최진영", "jinyoungchoi95@gmail.com");
+        Map<EntityColumn, Object> actual = new EntityClass<>(TestEntity.class).getFieldValues(givenObject);
+
+        assert actual != null;
+        assertAll(
+                () -> assertThat(actual).hasSize(2),
+                () -> assertThat(parseEntityColumnValue(actual, "id")).isEqualTo(1L),
+                () -> assertThat(parseEntityColumnValue(actual, "nick_name")).isEqualTo("최진영")
+        );
+    }
+
+    private Object parseEntityColumnValue(final Map<EntityColumn, Object> entityColumns, final String key) {
+        return entityColumns.entrySet()
+                .stream()
+                .filter(entry -> entry.getKey().getFieldName().equals(key))
+                .findAny()
+                .map(Map.Entry::getValue)
+                .get();
+    }
+
+
     @Entity
     @Table(name = "new_table")
     static class TableEntity {
@@ -64,6 +87,27 @@ class EntityClassTest {
 
         public NoConstructorEntity(Long id) {
             this.id = id;
+        }
+    }
+
+    @Entity
+    static class TestEntity {
+        @Id
+        private Long id;
+
+        @Column(name = "nick_name")
+        private String name;
+
+        @Transient
+        private String email;
+
+        public TestEntity() {
+        }
+
+        public TestEntity(Long id, String name, String email) {
+            this.id = id;
+            this.name = name;
+            this.email = email;
         }
     }
 }
